+++ packages/websocket-plugin/src/websocket-handler.ts
@@ -21,7 +21,7 @@ export class WebSocketHandler {
   ) {
     actions.pipe(filter(t => t instanceof ConnectWebSocket)).subscribe(event => socket.connect());
     actions.pipe(filter(t => t instanceof DisconnectWebSocket)).subscribe(event => socket.disconnect());
-    actions.pipe(filter(t => t instanceof SendWebSocketMessage)).subscribe(({ payload }) => socket.next(payload));
+    actions.pipe(filter(t => t instanceof SendWebSocketMessage)).subscribe(({ payload }) => socket.send(payload));
     socket.subscribe(msg => {
       const type = getValue(msg, msg[config.typeKey]);
       store.dispatch({