+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-# Master
+# 2.1.0-beta.4 4/3/18
 - Feature: Added `snapshot` method to store to get raw value
 - Fix: Return `T` from `selectSnapshot`
+- Fix: Minification issues #206
+- Fix: Websockets not emitting correctly
 
 # 2.1.0-beta.3 4/3/18
 - Fix: Revert devtools trigger, due to circular